+++ libraries/resource_cisco_ospf_vrf.rb
@@ -65,7 +65,7 @@ class Chef
       def router_id(arg=nil)
         set_or_return(:router_id, arg, :kind_of => String, :callbacks => {
           "must be a valid IPv4 address" => lambda {
-            |addr| IPAddress.valid_ipv4? addr
+            |addr| addr.empty? or IPAddress.valid_ipv4? addr
           }
         })
       end